Move train_evaluate_lstm_model(), to avoid circular dependency
